requirements.txt
requests
beautifulsoup4
feedparser
deep-translator
PyYAML
streamlit
pandas
